Module rustup_mock::clitools

A mock distribution server used by tests/cli-v1.rs and tests/cli-v2.rs

Structs

Config
The configuration used by the tests in this module

SanitizedOutput

Enums

Scenario

Statics

CROSS_ARCH1

CROSS_ARCH2

MULTI_ARCH1

Functions

change_dir

cmd

env

expect_err

expect_err_ex

expect_ok

expect_ok_ex

expect_stderr_ok

expect_stdout_ok

run

set_current_dist_date
Change the current distribution manifest to a particular date

setup
Run this to create the test environment containing multirust, and a mock dist server.

this_host_triple